webapp/pyproject.toml
[project]
name = "beryl3-webapp"
version = "0.1.0"
description = "Beryl, collection management system. Web application"
authors = [
    { name = "Mateusz Dubiel", email = "mdubiel@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "bleach>=6.2.0",
    "django>=5.2,<6.0",
    "django-allauth>=65.9.0",
    "django-browser-reload>=1.18.0",
    "django-environ>=0.12.0",
    "django-htmx>=1.23.1",
    "django-nanoid-field>=0.1.4",
    "django-post-office>=3.8.0",
    "django-prometheus>=2.3.1",
    "django-storages[google]>=1.14.4",
    "django-widget-tweaks>=1.5.0",
    "faker>=37.4.0",
    "google-cloud-logging>=3.12.0",
    "gunicorn>=23.0.0",
    "jsonschema>=4.25.1",
    "lucide>=1.1.3",
    "markdown>=3.9",
    "pillow>=11.2.1",
    "nudenet>=3.4.2",
    "psycopg2-binary>=2.9.10",
    "pylint-django>=2.6.1",
    "python-json-logger>=3.3.0",
    "python-logging-loki>=0.3.1",
    "pyyaml>=6.0",
    "requests>=2.32.0",
]

